Refactor construct into its own auto-found subcommand
